third_party/blink/web_tests/fast/shapes/parsing/parsing-shape-property-aliases-expected.txt

Test that CSS Shapes properties with the -webkit prefix alias the unprefixed properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getAliasCSSValue('shape-outside','shape-outside','circle(0px at 0% 0%)') is "circle(0px at 0% 0%)"
PASS getAliasCSSValue('shape-margin','shape-margin','10px') is "10px"
PASS getAliasCSSValue('shape-image-threshold','shape-image-threshold','0.5') is "0.5"
PASS successfullyParsed is true

TEST COMPLETE
